co-ordinator/requirements.txt
django==1.9.8
django-environ==0.4.0
djangorestframework==3.4.0
django-cors-headers

